third_party/blink/web_tests/http/tests/credentialmanager/credentialscontainer-create-with-resident-keys.html

<!DOCTYPE html>
<title>Credential Manager: End-to-end tests for create(), exercising creation of resident keys on virtual authenticators.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings_lite.js"></script>
<script src="/gen/mojo/public/mojom/base/time.mojom-lite.js"></script>
<script src="/gen/url/mojom/url.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/authenticator.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/virtual_authenticator.mojom-lite.js"></script>
<script src="resources/test-inputs.js"></script>
<script src="resources/virtual-navigator-credentials.js"></script>
<body>
<script>

if (document.location.host != "subdomain.example.test:8443") {
  document.location = "https://subdomain.example.test:8443/credentialmanager/credentialscontainer-create-with-resident-keys.html";
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
}

const credPropsTests = [
  {
    name: "U2F",
    authenticatorArgs: {
      protocol: blink.test.mojom.ClientToAuthenticatorProtocol.U2F,
    },
    expected: {
      discouraged: {
        success: true,
        hasRk: true,
        rk: false,
      },
      preferred: {
        success: true,
        hasRk: true,
        rk: false,
      },
      required: {
        success: false,
      },
    },
  },
  {
    name: "CTAP 2.0 without resident key support",
    authenticatorArgs: {
      protocol: blink.test.mojom.ClientToAuthenticatorProtocol.CTAP2,
      ctap2Version: blink.test.mojom.Ctap2Version.CTAP2_0,
      hasResidentKey: false,
    },
    expected: {
      discouraged: {
        success: true,
        hasRk: true,
        rk: false,
      },
      preferred: {
        success: true,
        hasRk: true,
        rk: false,
      },
      required: {
        success: false,
      },
    },
  },
  {
    name: "CTAP 2.0 with resident key support",
    authenticatorArgs: {
      protocol: blink.test.mojom.ClientToAuthenticatorProtocol.CTAP2,
      ctap2Version: blink.test.mojom.Ctap2Version.CTAP2_0,
      hasResidentKey: true,
    },
    expected: {
      discouraged: {
        success: true,
        // CTAP2.0 authenticators may treat all credentials as discoverable, thus Chrome omits 'rk' in this case.
        hasRk: false,
      },
      preferred: {
        success: true,
        hasRk: true,
        rk: true,
      },
      required: {
        success: true,
        hasRk: true,
        rk: true,
      },
    },
  },
  {
    name: "CTAP 2.1 without resident key support",
    authenticatorArgs: {
      protocol: blink.test.mojom.ClientToAuthenticatorProtocol.CTAP2,
      ctap2Version: blink.test.mojom.Ctap2Version.CTAP2_1,
      hasResidentKey: false,
    },
    expected: {
      discouraged: {
        success: true,
        hasRk: true,
        rk: false,
      },
      preferred: {
        success: true,
        hasRk: true,
        rk: false,
      },
      required: {
        success: false,
      },
    },
  },
  {
    name: "CTAP 2.1 with resident key support",
    authenticatorArgs: {
      protocol: blink.test.mojom.ClientToAuthenticatorProtocol.CTAP2,
      ctap2Version: blink.test.mojom.Ctap2Version.CTAP2_1,
      hasResidentKey: true,
    },
    expected: {
      discouraged: {
        success: true,
        hasRk: true,
        rk: false,
      },
      preferred: {
        success: true,
        hasRk: true,
        rk: true,
      },
      required: {
        success: true,
        hasRk: true,
        rk: true,
      },
    },
  },
];

for (const fixture of credPropsTests) {
  authenticatorSetup(fixture.name, () => {
    for (const rkRequirement of ["discouraged", "preferred", "required"]) {
      promise_test(async t => {
        var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
        customMakeCredOptions.extensions = {credProps: true};

        customMakeCredOptions.authenticatorSelection.residentKey = rkRequirement;
        const promise = navigator.credentials.create({publicKey: customMakeCredOptions});

        assert_true(rkRequirement in fixture.expected);
        const expected = fixture.expected[rkRequirement];
        assert_true('success' in expected);
        if (!expected.success) {
            return promise_rejects_dom(t, "NotAllowedError", promise);
        }

        const cred = await promise;
        assert_true('credProps' in cred.getClientExtensionResults());
        const credProps = cred.getClientExtensionResults().credProps;
        assert_equals('rk' in credProps, expected.hasRk, "hasRk");
        if (expected.hasRk) {
          assert_equals(credProps.rk, expected.rk, "rk");
        }
      }, fixture.name + ": navigator.credentials.create() with credProps extension, rk=" + rkRequirement);
    }
  }, fixture.authenticatorArgs);
}

</script>
